+++ culturemesh/__init__.py
@@ -18,4 +18,6 @@ import culturemesh.views
 # Register Blueprints
 
 from culturemesh.user_home.controllers import user_home
+from culturemesh.search.controllers import search
 app.register_blueprint(user_home, url_prefix='/home')
+app.register_blueprint(search, url_prefix='/search')
