+++ culturemesh/user_home/controllers.py
@@ -1,5 +1,4 @@
-from flask import Blueprint, request, render_template, redirect, url_for
-
+from flask import Blueprint, render_template
 from culturemesh.client import Client
 
 import flask_login
